Initial brainstorming
On Monday (Nov 1) the kids and I sat down and brainstormed some ideas
around what our game should be. You can see the short notes in
`Brainstorm.md` and `Rogeulike Brainstorming.md`.

In the end we settled on a bumblebee sim. The idea is to fly from flower
to flower in a random meadow avoiding dangers. Charlotte is starting on
the art, Audrey on the music and sound effects, and I'm getting a really
simple prototype engine ready.