Define the Bee hitbox and origin
Rework the Bee drawing code. Make a const RECT for the Bee sprite where
the Bee position is the origin. Also define a Bee hitbox, though it is
only displayed at the moment.

To make things easy, rotate the sprites. And to save space and
resolution, crop them to their bounding boxes.